Fixed compilation from MSYS
There was a bracket not guarding a return in engine.c MSVC accepted this, but GCC didn't seem to fancy it so much

Tested platforms:
- Windows 11, Visual Studio 2022, MSVC
- Windows 11, MSYS2, GCC

+++ include/rendimientos.h
@@ -16,15 +16,15 @@
 float TREL_EXPORT volumen_combustible(engine_t *engine);
 float TREL_EXPORT masa_combustible(engine_t* engine);
 float TREL_EXPORT Tiempo_quemado_combustion(engine_t* engine);
-float TREL_EXPORT Ab_combustion(engine_t* engine);
+double TREL_EXPORT Ab_combustion(engine_t* engine);
 float TREL_EXPORT br_combustion(engine_t* engine);
-float TREL_EXPORT mg_combustion(engine_t* engine);
-float TREL_EXPORT emp_teo(engine_t* engine);
+double TREL_EXPORT mg_combustion(engine_t* engine);
+double TREL_EXPORT emp_teo(engine_t* engine);
 float TREL_EXPORT emp_prom(float list[]);
 float TREL_EXPORT emp_max(float list[]);
 float TREL_EXPORT I_sp_teo(engine_t* engine);
 float TREL_EXPORT I_sp_esperado(float list[], engine_t *engine);
-float TREL_EXPORT I_tot_teo(engine_t* engine);
+double TREL_EXPORT I_tot_teo(engine_t* engine);
 float TREL_EXPORT I_tot_esp(float list[]);
 float TREL_EXPORT Delta_V_Esp(float list[]);
